+++ src/euroeval/dataset_configs/__init__.py
@@ -4,6 +4,7 @@ from ..data_models import DatasetConfig
 from ..languages import get_all_languages
 from ..tasks import SPEED
 from ..utils import load_custom_datasets_module
+from .bosnian import *  # noqa: F403
 from .bulgarian import *  # noqa: F403
 from .croatian import *  # noqa: F403
 from .czech import *  # noqa: F403