+++ src/scripts/create_lr_sum.py
@@ -15,7 +15,7 @@ from constants import MAX_NUM_CHARS_IN_ARTICLE, MIN_NUM_CHARS_IN_ARTICLE
 from datasets import Dataset, DatasetDict, Split, load_dataset
 from huggingface_hub import HfApi
 
-LANGUAGES = {"uk": "ukr", "sr": "srp"}
+LANGUAGES = {"uk": "ukr", "sr": "srp", "bs": "bos"}
 
 
 def main() -> None: