+++ src/scripts/create_multi_wiki_qa.py
@@ -50,6 +50,7 @@ LANGUAGES = [
     "sr",
     "sl",
     "hr",
+    "bs",
 ]